start paper 1 figure script
